feat: adding a review now takes username instead of user ID #19, fix: user languages and favourite genres now update correctly #23

+++ SocialBookAppDomain/Book.cs
@@ -5,7 +5,7 @@ public class Book
 {
     public int Id { get; set; }
     public string Title { get; set; } = string.Empty;
-    public int ISBNNumber { get; set; }
+    public long ISBNNumber { get; set; }
     public List<Author> Authors { get; set; } = new List<Author>();
     public DateTime Published { get; set; }
     public Genre Genre { get; set; }
